+++ scripts/_functions.sh
@@ -1,4 +1,6 @@
-#!/bin/bash -eu
+#!/bin/bash
+
+set -euxo pipefail
 
 function container() {
   docker-compose run --rm container $@