+++ package/CHANGELOG.md
@@ -23,3 +23,11 @@
 "@types/react-redux",
 "@testing-library/dom"
 ```
+
+## 4.0.0
+
+-   removed cleanup from examples since it happens by default
+
+### Breaking Changes
+
+The updateStateWithDispatch method is no longer async. If you need to wait for async code to run after a dispatch consider using `await new Promise(setImmediate);` to flush promises.